Issue 3986 - AgentImageCleanup: Agent handles new type of NMP
